ein paar unnoetige Klassen geloescht
