--- Code/src/view/weltraum/WeltraumSicht1.java
@@ -1,20 +1,0 @@
-package view.weltraum;
-
-import javafx.scene.layout.AnchorPane;
-import map.Sektion;
-
-/**
- * In dieser Sichtweise kann der Spieler jede Deteis seiner umgebung erkennen. Die einzelnen Raumschiffe, 
- * Planeten, Monde und weiteres
- */
-public class WeltraumSicht1 extends AnchorPane
-{
-	/**
-	 * die Sektion in der sich der Spieler befindet muss hier mit uebergeben werden
-	 * @param sektion
-	 */
-	public WeltraumSicht1(Sektion sektion)
-	{
-		
-	}
-}
