--- Code/src/view/weltraum/WeltraumSicht2.java
@@ -1,21 +1,0 @@
-package view.weltraum;
-
-import map.Sektion;
-
-/**
- * In dieser Sicht, hat der Spieler ein ueberblick ueber die Planeten und deren Monde. Raumschiffe werden mit 
- * Symbolen gekenzeichnet. 
- * 
- * @author Dennis
- */
-public class WeltraumSicht2 
-{
-	/**
-	 * die Sektion in der sich der Spieler befindet muss hier mit uebergeben werden
-	 * @param sektion
-	 */
-	public WeltraumSicht2(Sektion sektion)
-	{
-		
-	}
-}
