--- Code/src/view/weltraum/WeltraumSicht4.java
@@ -1,22 +1,0 @@
-package view.weltraum;
-
-import map.Sektion;
-
-/**
- * In dieser Sicht werden mehrere Sektionen angezeigt.
- * @author Dennis
- * 
- * TODO weitere Sichtweisen moeglich
- *
- */
-public class WeltraumSicht4 
-{
-	/**
-	 * die Sektion in der sich der Spieler befindet muss hier mit uebergeben werden
-	 * @param sektion
-	 */
-	public WeltraumSicht4(Sektion sektion)
-	{
-		
-	}
-}
